+++ parser/sql/statement/core/src/test/java/org/apache/shardingsphere/sql/parser/statement/core/util/SubqueryExtractUtilsTest.java
@@ -72,9 +72,10 @@ class SubqueryExtractUtilsTest {
         SubqueryExpressionSegment right = new SubqueryExpressionSegment(new SubquerySegment(51, 100, subquerySelectStatement, ""));
         WhereSegment whereSegment = new WhereSegment(34, 100, new BinaryOperationExpression(40, 100, left, right, "=", "order_id = (SELECT order_id FROM t_order WHERE status = 'OK')"));
         when(selectStatement.getWhere()).thenReturn(Optional.of(whereSegment));
-        Collection<SubquerySegment> actual = SubqueryExtractUtils.getSubquerySegments(selectStatement);
+        Collection<SubquerySegment> actual = SubqueryExtractUtils.getSubquerySegments(selectStatement, true);
         assertThat(actual.size(), is(1));
         assertThat(actual.iterator().next(), is(right.getSubquery()));
+        assertThat(SubqueryExtractUtils.getSubquerySegments(selectStatement, false).size(), is(1));
     }
     
     @Test
@@ -89,9 +90,10 @@ class SubqueryExtractUtilsTest {
         ProjectionsSegment projections = new ProjectionsSegment(7, 79);
         when(selectStatement.getProjections()).thenReturn(projections);
         projections.getProjections().add(subqueryProjectionSegment);
-        Collection<SubquerySegment> actual = SubqueryExtractUtils.getSubquerySegments(selectStatement);
+        Collection<SubquerySegment> actual = SubqueryExtractUtils.getSubquerySegments(selectStatement, true);
         assertThat(actual.size(), is(1));
         assertThat(actual.iterator().next(), is(subquerySegment));
+        assertThat(SubqueryExtractUtils.getSubquerySegments(selectStatement, false).size(), is(1));
     }
     
     @Test
@@ -110,9 +112,10 @@ class SubqueryExtractUtilsTest {
         projections.getProjections().add(new ColumnProjectionSegment(new ColumnSegment(7, 16, new IdentifierValue("order_id"))));
         SubqueryTableSegment subqueryTableSegment = new SubqueryTableSegment(0, 0, new SubquerySegment(23, 71, subquery, ""));
         when(selectStatement.getFrom()).thenReturn(Optional.of(subqueryTableSegment));
-        Collection<SubquerySegment> actual = SubqueryExtractUtils.getSubquerySegments(selectStatement);
+        Collection<SubquerySegment> actual = SubqueryExtractUtils.getSubquerySegments(selectStatement, true);
         assertThat(actual.size(), is(1));
         assertThat(actual.iterator().next(), is(subqueryTableSegment.getSubquery()));
+        assertThat(SubqueryExtractUtils.getSubquerySegments(selectStatement, false).size(), is(1));
     }
     
     @Test
@@ -154,11 +157,12 @@ class SubqueryExtractUtilsTest {
         from.setLeft(leftSubquerySegment);
         from.setRight(rightSubquerySegment);
         when(selectStatement.getFrom()).thenReturn(Optional.of(from));
-        Collection<SubquerySegment> actual = SubqueryExtractUtils.getSubquerySegments(selectStatement);
+        Collection<SubquerySegment> actual = SubqueryExtractUtils.getSubquerySegments(selectStatement, true);
         assertThat(actual.size(), is(2));
         Iterator<SubquerySegment> iterator = actual.iterator();
         assertThat(iterator.next(), is(leftSubquerySegment.getSubquery()));
         assertThat(iterator.next(), is(rightSubquerySegment.getSubquery()));
+        assertThat(SubqueryExtractUtils.getSubquerySegments(selectStatement, false).size(), is(2));
     }
     
     @Test
@@ -166,8 +170,8 @@ class SubqueryExtractUtilsTest {
         SelectStatement selectStatement = mock(SelectStatement.class);
         SubquerySegment subquerySelect = createSubquerySegment();
         when(selectStatement.getFrom()).thenReturn(Optional.of(new SubqueryTableSegment(0, 0, subquerySelect)));
-        Collection<SubquerySegment> actual = SubqueryExtractUtils.getSubquerySegments(selectStatement);
-        assertThat(actual.size(), is(2));
+        assertThat(SubqueryExtractUtils.getSubquerySegments(selectStatement, true).size(), is(2));
+        assertThat(SubqueryExtractUtils.getSubquerySegments(selectStatement, false).size(), is(1));
     }
     
     @Test
@@ -176,8 +180,8 @@ class SubqueryExtractUtilsTest {
         SubquerySegment left = new SubquerySegment(0, 0, mock(SelectStatement.class), "");
         SubquerySegment right = createSubquerySegment();
         when(selectStatement.getCombine()).thenReturn(Optional.of(new CombineSegment(0, 0, left, CombineType.UNION, right)));
-        Collection<SubquerySegment> actual = SubqueryExtractUtils.getSubquerySegments(selectStatement);
-        assertThat(actual.size(), is(3));
+        assertThat(SubqueryExtractUtils.getSubquerySegments(selectStatement, true).size(), is(3));
+        assertThat(SubqueryExtractUtils.getSubquerySegments(selectStatement, false).size(), is(2));
     }
     
     private SubquerySegment createSubquerySegment() {
@@ -197,7 +201,7 @@ class SubqueryExtractUtilsTest {
         functionSegment.getParameters().add(new SubqueryExpressionSegment(new SubquerySegment(0, 0, mock(SelectStatement.class), "")));
         ExpressionProjectionSegment expressionProjectionSegment = new ExpressionProjectionSegment(0, 0, "", functionSegment);
         projections.getProjections().add(expressionProjectionSegment);
-        Collection<SubquerySegment> actual = SubqueryExtractUtils.getSubquerySegments(selectStatement);
-        assertThat(actual.size(), is(1));
+        assertThat(SubqueryExtractUtils.getSubquerySegments(selectStatement, true).size(), is(1));
+        assertThat(SubqueryExtractUtils.getSubquerySegments(selectStatement, false).size(), is(1));
     }
 }